Simplify Callable API
