+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/CallableApi.java
@@ -18,7 +18,14 @@
 
 package eu.fasten.analyzer.restapiplugin.mvn.api;
 
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
+import eu.fasten.analyzer.restapiplugin.mvn.LazyIngestionProvider;
 import eu.fasten.analyzer.restapiplugin.mvn.RestApplication;
+import eu.fasten.core.maven.data.PackageVersionNotFoundException;
+import org.json.JSONObject;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,16 +34,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import java.io.IOException;
 import java.util.List;
 
 @RestController
 public class CallableApi {
-
-    private final CallableApiService service;
-
-    public CallableApi(CallableApiService service) {
-        this.service = service;
-    }
 
     @GetMapping(value = "/mvn/packages/{pkg}/{pkg_ver}/callables", produces = MediaType.APPLICATION_JSON_VALUE)
     ResponseEntity<String> getPackageCallables(@PathVariable("pkg") String package_name,
@@ -45,20 +47,54 @@ public class CallableApi {
                                                @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) int limit,
                                                @RequestParam(required = false) String artifactRepository,
                                                @RequestParam(required = false) Long releaseDate) {
-        return service.getPackageCallables(package_name, package_version, offset, limit, artifactRepository, releaseDate);
+        String result;
+        try {
+            result = KnowledgeBaseConnector.kbDao.getPackageCallables(
+                    package_name, package_version, offset, limit);
+        } catch (PackageVersionNotFoundException e) {
+            try {
+                LazyIngestionProvider.ingestArtifactIfNecessary(package_name, package_version, artifactRepository, releaseDate);
+            } catch (IllegalArgumentException ex) {
+                return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
+            } catch (IOException ex) {
+                return new ResponseEntity<>(ex.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            return new ResponseEntity<>("Package version not found, but should be processed soon. Try again later", HttpStatus.CREATED);
+        }
+        result = result.replace("\\/", "/");
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @PostMapping(value = "/mvn/packages/{pkg}/{pkg_ver}/callable/metadata", produces = MediaType.APPLICATION_JSON_VALUE)
     ResponseEntity<String> getCallableMetadata(@PathVariable("pkg") String package_name,
                                                @PathVariable("pkg_ver") String package_version,
                                                @RequestBody String fasten_uri,
-                                               @RequestParam(value = "artifactRepository", required = false) String artifactRepo,
+                                               @RequestParam(value = "artifactRepository", required = false) String artifactRepository,
                                                @RequestParam(required = false) Long releaseDate) {
-        return service.getCallableMetadata(package_name, package_version, fasten_uri, artifactRepo, releaseDate);
+        String result = KnowledgeBaseConnector.kbDao.getCallableMetadata(
+                package_name, package_version, fasten_uri);
+        if (result == null) {
+            try {
+                LazyIngestionProvider.ingestArtifactIfNecessary(package_name, package_version, artifactRepository, releaseDate);
+            } catch (IllegalArgumentException | IllegalStateException | IOException ex) {
+                return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
+            }
+            return new ResponseEntity<>("Package version not found, but should be processed soon. Try again later", HttpStatus.CREATED);
+        }
+        result = result.replace("\\/", "/");
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
     @PostMapping(value = "/callables", produces = MediaType.APPLICATION_JSON_VALUE)
     ResponseEntity<String> getCallables(@RequestBody List<Long> callableIDs) {
-        return service.getCallables(callableIDs);
+
+        var callablesMap = KnowledgeBaseConnector.kbDao.getCallables(callableIDs);
+        var json = new JSONObject();
+        for (var id : callableIDs) {
+            json.put(String.valueOf(id), callablesMap.get(id));
+        }
+        var result = json.toString();
+        result = result.replace("\\/", "/");
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
